+++ CHANGELOG.md
@@ -3,9 +3,7 @@
 - Added syntax support for subscription queries (they compile/no error but do not execute or work)
 - Removed support for older syntax of complex queries that is not GQL standard
 - Refactored `GraphQLVistor` and friends to make it easier to follow what is happening (I hope). See CONTRIBUTING.md for some notes
-
-- Can now mark fields or enum values as deprecated
-- Schema generation with output deprecated fields marked with `@deprecated(reason)`
+- From my testing the compiling and expression building is 15-20% faster than before (still network and or the DB calls are the largest)
 
 # 0.50.1
 - Name all `ParameterExpression`s as EF 3.1 expects a name (can throw an error)